+++ HiraethEngine3/HiraethEngine3/src/hm/colour.hpp
@@ -3,16 +3,16 @@
 namespace hm {
     
     struct colour {
-        int r, g, b, a; // rgba in range of [0:255] 
+        uint8_t r, g, b, a; // rgba in range of [0:255] 
         float i; // intensity of the colour [0:inf]
         
         colour() : r(0), g(0), b(0), a(0), i(1) {};
-        colour(int v) : r(v), g(v), b(v), a(255), i(1.f) {};
-        colour(int v, float i) : r(v), g(v), b(v), a(255), i(i) {};
-        colour(int r, int g, int b) : r(r), g(g), b(b), a(255), i(1) {};
-        colour(int r, int g, int b, int a) : r(r), g(g), b(b), a(a), i(1.f) {};
-        colour(int r, int g, int b, float i) : r(r), g(g), b(b), a(255), i(i) {};
-        colour(int r, int g, int b, int a, float i) : r(r), g(g), b(b), a(a), i(i) {};
+        colour(uint8_t v) : r(v), g(v), b(v), a(255), i(1.f) {};
+        colour(uint8_t v, float i) : r(v), g(v), b(v), a(255), i(i) {};
+        colour(uint8_t r, uint8_t g, uint8_t b) : r(r), g(g), b(b), a(255), i(1) {};
+        colour(uint8_t r, uint8_t g, uint8_t b, uint8_t a) : r(r), g(g), b(b), a(a), i(1.f) {};
+        colour(uint8_t r, uint8_t g, uint8_t b, float i) : r(r), g(g), b(b), a(255), i(i) {};
+        colour(uint8_t r, uint8_t g, uint8_t b, uint8_t a, float i) : r(r), g(g), b(b), a(a), i(i) {};
     };
     
     static float getR(const colour* col) { return (col->r * col->i) / 255.f; };
